Cleanup post refactor
Keeps as much of the common elements of XDP and TC dispatcher handling
as possible together to (hopefully) reduce complexity.

+++ bpfd/src/command.rs
@@ -2,21 +2,20 @@
 // Copyright Authors of bpfd
 
 //! Commands between the RPC thread and the BPF thread
-use std::{fmt, fs, io::BufReader, str::FromStr};
+use std::{fmt, fs, io::BufReader, path::PathBuf, str::FromStr};
 
-use bpfd_api::{util::directories::RTDIR_PROGRAMS, ParseError};
+use bpfd_api::{
+    util::directories::{RTDIR_FS, RTDIR_PROGRAMS},
+    ParseError,
+};
 use serde::{Deserialize, Serialize};
 use tokio::sync::oneshot;
 use uuid::Uuid;
 
-use crate::errors::BpfdError;
-
-pub(crate) const DEFAULT_XDP_PROCEED_ON_PASS: i32 = 2;
-pub(crate) const DEFAULT_XDP_PROCEED_ON_DISPATCHER_RETURN: i32 = 31;
-// Default is Pass and DispatcherReturn
-pub(crate) const DEFAULT_XDP_ACTIONS_MAP: u32 =
-    1 << DEFAULT_XDP_PROCEED_ON_PASS as u32 | 1 << DEFAULT_XDP_PROCEED_ON_DISPATCHER_RETURN as u32;
-pub(crate) const DEFAULT_ACTIONS_MAP_TC: u32 = 1 << 3; // TC_ACT_PIPE;
+use crate::{
+    errors::BpfdError,
+    multiprog::{DispatcherId, DispatcherInfo, TC_ACT_PIPE, XDP_DISPATCHER_RET, XDP_PASS},
+};
 
 /// Provided by the requester and used by the manager task to send
 /// the command response back to the requester.
@@ -126,7 +125,7 @@ impl std::fmt::Display for Direction {
 pub struct NetworkMultiAttach {
     pub(crate) iface: String,
     pub(crate) priority: i32,
-    pub(crate) proceed_on: Vec<i32>,
+    pub(crate) proceed_on: ProceedOn,
     pub(crate) position: i32,
 }
 
@@ -140,65 +139,167 @@ pub(crate) struct ProgramInfo {
     pub(crate) attach_type: AttachType,
 }
 
-#[derive(Debug, Eq, Ord, PartialEq, PartialOrd, Serialize, Deserialize)]
+#[derive(Debug, Eq, Ord, PartialEq, PartialOrd, Serialize, Deserialize, Clone)]
 pub(crate) struct Metadata {
     pub(crate) priority: i32,
     pub(crate) name: String,
     pub(crate) attached: bool,
 }
 
-#[derive(Serialize, Deserialize)]
-pub(crate) enum Program {
-    Xdp(ProgramData, NetworkMultiAttachInfo),
-    Tracepoint(ProgramData, String),
-    Tc(ProgramData, NetworkMultiAttachInfo, Direction),
+impl Metadata {
+    pub(crate) fn new(priority: i32, name: String) -> Self {
+        Metadata {
+            priority,
+            name,
+            attached: false,
+        }
+    }
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone)]
+pub(crate) enum Program {
+    Xdp(XdpProgram),
+    Tracepoint(TracepointProgram),
+    Tc(TcProgram),
+}
+
+#[derive(Serialize, Deserialize, Clone)]
+pub(crate) struct XdpProgram {
+    pub(crate) data: ProgramData,
+    pub(crate) info: NetworkMultiAttachInfo,
+}
+
+impl XdpProgram {
+    pub(crate) fn new(data: ProgramData, info: NetworkMultiAttachInfo) -> Self {
+        Self { data, info }
+    }
+}
+
+#[derive(Serialize, Deserialize, Clone)]
+pub(crate) struct TcProgram {
+    pub(crate) data: ProgramData,
+    pub(crate) info: NetworkMultiAttachInfo,
+    pub(crate) direction: Direction,
+}
+#[derive(Serialize, Deserialize, Clone)]
+pub(crate) struct TracepointProgram {
+    pub(crate) data: ProgramData,
+    pub(crate) info: String,
+}
+
+#[derive(Serialize, Deserialize, Clone)]
 pub(crate) struct ProgramData {
     pub(crate) path: String,
     pub(crate) section_name: String,
     pub(crate) owner: String,
 }
 
-#[derive(Serialize, Deserialize)]
+impl ProgramData {
+    pub(crate) fn new(path: String, section_name: String, owner: String) -> Self {
+        ProgramData {
+            path,
+            section_name,
+            owner,
+        }
+    }
+}
+
+#[derive(Serialize, Deserialize, Clone)]
 pub(crate) struct NetworkMultiAttachInfo {
     pub(crate) if_name: String,
     pub(crate) if_index: u32,
     #[serde(skip)]
     pub(crate) current_position: Option<usize>,
     pub(crate) metadata: Metadata,
-    pub(crate) proceed_on: Vec<i32>,
+    pub(crate) proceed_on: ProceedOn,
 }
 
 impl NetworkMultiAttachInfo {
-    pub(crate) fn proceed_on_mask(&self) -> u32 {
+    pub(crate) fn new(
+        if_name: String,
+        if_index: u32,
+        metadata: Metadata,
+        proceed_on: ProceedOn,
+    ) -> Self {
+        NetworkMultiAttachInfo {
+            if_name,
+            if_index,
+            current_position: None,
+            metadata,
+            proceed_on,
+        }
+    }
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub(crate) struct ProceedOn(pub(crate) Vec<i32>);
+impl ProceedOn {
+    pub fn default_xdp() -> Self {
+        ProceedOn(vec![XDP_PASS, XDP_DISPATCHER_RET])
+    }
+
+    pub fn default_tc() -> Self {
+        ProceedOn(vec![TC_ACT_PIPE])
+    }
+
+    pub(crate) fn mask(&self) -> u32 {
         let mut proceed_on_mask: u32 = 0;
-        if self.proceed_on.is_empty() {
-            proceed_on_mask = DEFAULT_XDP_ACTIONS_MAP;
-        } else {
-            for action in self.proceed_on.clone().into_iter() {
-                proceed_on_mask |= 1 << action;
-            }
+        for action in self.0.clone().into_iter() {
+            proceed_on_mask |= 1 << action;
         }
         proceed_on_mask
     }
 }
 
 impl Program {
+    pub(crate) fn dispatcher_id(&self) -> Option<DispatcherId> {
+        match self {
+            Program::Xdp(p) => Some(DispatcherId::Xdp(DispatcherInfo(p.info.if_index, None))),
+            Program::Tc(p) => Some(DispatcherId::Tc(DispatcherInfo(
+                p.info.if_index,
+                Some(p.direction),
+            ))),
+            _ => None,
+        }
+    }
+
+    pub(crate) fn data(&self) -> &ProgramData {
+        match self {
+            Program::Xdp(p) => &p.data,
+            Program::Tracepoint(p) => &p.data,
+            Program::Tc(p) => &p.data,
+        }
+    }
+
+    pub(crate) fn metadata(&self) -> Option<&Metadata> {
+        match self {
+            Program::Xdp(p) => Some(&p.info.metadata),
+            Program::Tracepoint(_) => None,
+            Program::Tc(p) => Some(&p.info.metadata),
+        }
+    }
+
     pub(crate) fn owner(&self) -> &String {
         match self {
-            Program::Xdp(d, _) => &d.owner,
-            Program::Tracepoint(d, _) => &d.owner,
-            Program::Tc(d, _, _) => &d.owner,
+            Program::Xdp(p) => &p.data.owner,
+            Program::Tracepoint(p) => &p.data.owner,
+            Program::Tc(p) => &p.data.owner,
         }
     }
 
     pub(crate) fn set_attached(&mut self) {
         match self {
-            Program::Xdp(_, m) => m.metadata.attached = true,
-            Program::Tc(_, m, _) => m.metadata.attached = true,
-            Program::Tracepoint(_, _) => (),
+            Program::Xdp(p) => p.info.metadata.attached = true,
+            Program::Tc(p) => p.info.metadata.attached = true,
+            Program::Tracepoint(_) => (),
+        }
+    }
+
+    pub(crate) fn set_position(&mut self, pos: Option<usize>) {
+        match self {
+            Program::Xdp(p) => p.info.current_position = pos,
+            Program::Tc(p) => p.info.current_position = pos,
+            Program::Tracepoint(_) => (),
         }
     }
 
@@ -211,11 +312,17 @@ impl Program {
     pub(crate) fn delete(&self, uuid: Uuid) -> Result<(), anyhow::Error> {
         let path = format!("{RTDIR_PROGRAMS}/{uuid}");
         fs::remove_file(path)?;
-        let path = format!("/var/run/bpfd/fs/prog_{uuid}");
-        fs::remove_file(path)?;
+
+        let path = format!("{RTDIR_FS}/prog_{uuid}");
+        if PathBuf::from(&path).exists() {
+            fs::remove_file(path)?;
+        }
+        let path = format!("{RTDIR_FS}/prog_{uuid}_link");
+        if PathBuf::from(&path).exists() {
+            fs::remove_file(path)?;
+        }
         let path = format!("/var/run/bpfd/fs/maps/{uuid}");
         fs::remove_dir_all(path)?;
-
         Ok(())
     }
 
@@ -225,5 +332,37 @@ impl Program {
         let reader = BufReader::new(file);
         let prog = serde_json::from_reader(reader)?;
         Ok(prog)
+    }
+
+    pub(crate) fn kind(&self) -> ProgramType {
+        match self {
+            Program::Xdp(_) => ProgramType::Xdp,
+            Program::Tracepoint(_) => ProgramType::Tracepoint,
+            Program::Tc(_) => ProgramType::Tc,
+        }
+    }
+
+    pub(crate) fn if_index(&self) -> Option<u32> {
+        match self {
+            Program::Xdp(p) => Some(p.info.if_index),
+            Program::Tracepoint(_) => None,
+            Program::Tc(p) => Some(p.info.if_index),
+        }
+    }
+
+    pub(crate) fn if_name(&self) -> Option<String> {
+        match self {
+            Program::Xdp(p) => Some(p.info.if_name.clone()),
+            Program::Tracepoint(_) => None,
+            Program::Tc(p) => Some(p.info.if_name.clone()),
+        }
+    }
+
+    pub(crate) fn direction(&self) -> Option<Direction> {
+        match self {
+            Program::Xdp(_) => None,
+            Program::Tracepoint(_) => None,
+            Program::Tc(p) => Some(p.direction),
+        }
     }
 }
